Added Vue.js UI for notifications
